+++ modules/commands/includes/CommandSpec.hpp
@@ -4,9 +4,9 @@
 #include <sstream>
 
 #include "ft_irc.hpp"
-#include "parsing/includes/CommandElement.hpp"
-#include "parsing/includes/CommandExecutor.hpp"
-#include "parsing/includes/exception/TooFewArgumentsException.hpp"
+#include "commands/includes/CommandElement.hpp"
+#include "commands/includes/CommandExecutor.hpp"
+#include "commands/includes/exception/TooFewArgumentsException.hpp"
 
 class CommandSpec {
 private:
